requirements.txt
# FastAPI
fastapi==0.114.0
uvicorn[standard]==0.23.2

# Streamlit
streamlit==1.28.1

# Database / ORM
SQLAlchemy==2.0.20
mysql-connector-python==8.1.0

# Data & plotting
pandas==2.1.1
numpy==1.27.2
matplotlib==3.8.0
seaborn==0.12.3
plotly==5.16.1

# Machine Learning
scikit-learn==1.3.1
tensorflow==2.13.0
joblib==1.3.2

# Optional / utility
python-dotenv==1.0.0
